change: added scribbles suggestions and fixed bug
