+++ scripts/cat/appearance_utility.py
@@ -122,60 +122,6 @@ def plural_acc_names(accessory ,plural, singular):
     elif singular is True and plural is False:
         return acc_display
 
-def accessory_display_name(accessory):
-    if not accessory:
-        return ''
-    accessory = accessory.lower()
-    acc_display = accessory
-    if accessory != None:
-        if accessory in collars:
-            collar_color = None
-            if accessory.startswith('crimson'):
-                collar_color = 'crimson'
-            elif accessory.startswith('blue'):
-                collar_color = 'blue'
-            elif accessory.startswith('yellow'):
-                collar_color = 'yellow'
-            elif accessory.startswith('cyan'):
-                collar_color = 'cyan'
-            elif accessory.startswith('red'):
-                collar_color = 'red'
-            elif accessory.startswith('lime'):
-                collar_color = 'lime'
-            elif accessory.startswith('green'):
-                collar_color = 'green'
-            elif accessory.startswith('rainbow'):
-                collar_color = 'rainbow'
-            elif accessory.startswith('black'):
-                collar_color = 'black'
-            elif accessory.startswith('spikes'):
-                collar_color = 'spiky'
-            elif accessory.startswith('pink'):
-                collar_color = 'pink'
-            elif accessory.startswith('purple'):
-                collar_color = 'purple'
-            elif accessory.startswith('multi'):
-                collar_color = 'multi'
-            if accessory.endswith('bow') and not accessory == 'rainbow':
-                acc_display = collar_color + ' bow'
-            elif accessory.endswith('bell'):
-                acc_display = collar_color + ' bell collar'
-            else:
-                acc_display = collar_color + ' collar'
-
-    elif accessory in wild_accessories:
-        if acc_display == 'blue feathers':
-            acc_display = 'crow feathers'
-        elif acc_display == 'red feathers':
-            acc_display = 'cardinal feathers'
-        else:
-            acc_display = acc_display
-    else:
-        acc_display = acc_display
-    if accessory == None:
-        acc_display = None
-    return acc_display
-
 # ---------------------------------------------------------------------------- #
 #                                init functions                                #
 # ---------------------------------------------------------------------------- #
